+++ examples/ethernet_networks/update_resource.tf
@@ -3,7 +3,7 @@ provider "oneview" {
         ov_password =   "${var.password}"
         ov_endpoint =   "${var.endpoint}"
         ov_sslverify =  "${var.ssl_enabled}"
-        ov_apiversion = 2600
+        ov_apiversion = 2800
         ov_ifmatch = "*"
 
 }
@@ -13,4 +13,16 @@ resource "oneview_ethernet_network" "ethernetnetwork" {
 	name = "TestEthNetwork_terraform_Rename"
 	type = "ethernet-networkV4"
 	vlan_id = 100
+}
+
+resource "oneview_ethernet_network" "ethernetnetwork_1" {
+	name = "Auto-Ethernet-1"
+	type = "ethernet-networkV4"
+	vlan_id = 101
+}
+
+resource "oneview_ethernet_network" "ethernetnetwork_2" {
+	name = "Auto-Ethernet-2"
+	type = "ethernet-networkV4"
+	vlan_id = 102
 }