+++ examples/storage_systems/data_source.tf
@@ -3,13 +3,13 @@ provider "oneview" {
   ov_password   = "${var.password}"
   ov_endpoint   = "${var.endpoint}"
   ov_sslverify  = "${var.ssl_enabled}"
-  ov_apiversion = 2600
+  ov_apiversion = 2800
   ov_ifmatch    = "*"
 }
 
 # Extracting Storage System
 data "oneview_storage_system" "ss_inst" {
-   name = "ThreePAR-2"
+   name = "ThreePAR-1"
 }
 
 output "oneview_storage_system_value" {