+++ examples/storage_systems/main.tf
@@ -3,30 +3,30 @@ provider "oneview" {
   ov_password   = "${var.password}"
   ov_endpoint   = "${var.endpoint}"
   ov_sslverify  = "${var.ssl_enabled}"
-  ov_apiversion = 2600
+  ov_apiversion = 2800
   ov_ifmatch    = "*"
 }
 
 variable "hostname" {
   type    = "string"
-  default = "<storage_system_ip>"
+  default = "172.18.11.11" //"<storage_system_ip>"
 }
 variable "ss_username" {
   type    = "string"
-  default = "<storage_system_username>"
+  default = "dcs" //"<storage_system_username>"
 }
 variable "ss_password" {
   type    = "string"
-  default = "<storage_system_password>"
+  default = "dcs" //"<storage_system_password>"
 }
 variable "ss_family" {
   type    = "string"
-  default = "<storage_system_family>"
+  default = "StoreServ" //"<storage_system_family>"
 }
 
 # Extracting Server Certificate
 data "oneview_server_certificate" "sc" {
-  remote_ip = "${var.hostname}"
+  remote_ip = "172.18.11.12" //"${var.hostname}"
 }
 
 # Importing Server Certificate for adding storage system
@@ -34,7 +34,7 @@ resource "oneview_server_certificate" "ServerCertificate" {
     certificate_details = [{
                         base64_data="${data.oneview_server_certificate.sc.certificate_details.0.base64_data}"
                         type="CertificateDetailV2"
-                        alias_name = "TestServerCertificate"
+                        alias_name = "StorageSystemCertificate"
                         }]
 }