+++ examples/storage_systems/update_resource.tf
@@ -3,30 +3,30 @@ provider "oneview" {
   ov_password   = "${var.password}"
   ov_endpoint   = "${var.endpoint}"
   ov_sslverify  = "${var.ssl_enabled}"
-  ov_apiversion = 2600
+  ov_apiversion = 2800
   ov_ifmatch    = "*"
 }
 
 variable "hostname" {
   type    = "string"
-  default = "<storage_system_ip>"
+  default = "172.18.11.11" //"<storage_system_ip>"
 }
 variable "ss_username" {
   type    = "string"
-  default = "<storage_system_username>"
+  default = "dcs" //"<storage_system_username>"
 }
 variable "ss_password" {
   type    = "string"
-  default = "<storage_system_password>"
+  default = "dcs" //"<storage_system_password>"
 }
 variable "ss_family" {
   type    = "string"
-  default = "<storage_system_family>"
+  default = "StoreServ" //"<storage_system_family>"
 }
 
 # Extracting Storage System
 data "oneview_storage_system" "ss_inst" {
-   name = "ThreePAR-2"
+   name = "ThreePAR-1"
 }
 
 # Updates the resource created from main.tf 
@@ -39,7 +39,7 @@ resource "oneview_storage_system" "ss_inst" {
   ]
 
   hostname =  "${var.hostname}"
-  name     = "ThreePAR-2"
+  name     = "ThreePAR-1"
 
   storage_system_device_specific_attributes = {
     managed_domain = "TestDomain"