+++ examples/volumes/main.tf
@@ -3,7 +3,7 @@ provider "oneview" {
   ov_password   = "${var.password}"
   ov_endpoint   = "${var.endpoint}"
   ov_sslverify  = "${var.ssl_enabled}"
-  ov_apiversion = 2600
+  ov_apiversion = 2800
   ov_ifmatch    = "*"
 }
 
@@ -15,6 +15,7 @@ data "oneview_scope" "scope_obj" {
 data "oneview_storage_pool" "st_pool" {
         name = "CPG-SSD-AO"
 }
+
 data "oneview_storage_volume_template" "st_vt" {
         name = "RenameDemoStorageTemplate"
 }